feature: combine the two files into one
I'm not 100% sure what the issue was, but it's working fine after turning them into one file
